# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "yy_easy_aiapi — AI 代理网关 API",
    "description": "轻量级 AI 代理网关，提供统一的 AI 接口调用、渠道管理、Token 鉴权与计费能力。\n\n## 鉴权方式\n\n### 1. TokenAuth（AI 代理鉴权）\n用于调用 AI 聊天补全接口。在请求头中添加：\n```\nAuthorization: Bearer {系统生成的Token}\n```\n\n### 2. ApiTokenAuth（管理 API 鉴权）\n用于调用所有渠道管理、Token 管理、积分操作、流水查询、计费设置接口。在请求头中添加：\n```\nAuthorization: Bearer {api_token}\n```\n\n## 统一响应格式\n所有 API 响应均为标准 JSON 格式：\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"error\": null,\n  \"message\": null\n}\n```\n\n错误响应：\n```json\n{\n  \"success\": false,\n  \"data\": null,\n  \"error\": {\n    \"code\": 401,\n    \"message\": \"错误描述\"\n  },\n  \"message\": \"错误描述\"\n}\n```\n\n## 计费公式\n**最终价格 = 渠道价格 × 倍率 × 100**\n\n其中：\n- 渠道价格：各渠道中配置的单价（元/单位）\n- 倍率：在计费设置中配置的全局倍率（默认 1.0）\n- 最终价格以积分为单位（1 元 = 100 积分）",
    "version": "1.0.0",
    "contact": {
      "name": "yy_easy_aiapi"
    }
  },
  "servers": [
    {
      "url": "/yy_easy_aiapi",
      "description": "yy_easy_aiapi API 基础路径（部署时需包含此前缀）"
    }
  ],
...
  "paths": {
...
    "/api/channels": {
      "get": {
        "tags": ["渠道管理"],
        "summary": "获取渠道列表",
        "description": "获取所有已配置的 AI 渠道列表。每个渠道代表一个上游 AI 服务提供商。\n\n### 响应示例\n```json\n{\n  \"success\": true,\n  \"data\": [\n    {\n      \"id\": 1,\n      \"name\": \"gpt-4o\",\n      \"api_standard\": \"openai\",\n      \"price_input\": \"0.01\",\n      \"price_cached\": \"0.005\",\n      \"price_output\": \"0.03\",\n      \"rate_input\": \"1.0\",\n      \"rate_cached\": \"1.0\",\n      \"rate_output\": \"1.0\",\n      \"upstream_url\": \"https://api.openai.com/v1\",\n      \"api_key\": \"sk-xxx\",\n      \"created_at\": \"2025-01-01T00:00:00\",\n      \"updated_at\": \"2025-01-01T00:00:00\"\n    }\n  ]\n}\n```",
        "operationId": "listChannels",
        "security": [{ "ApiTokenAuth": [] }],
        "responses": {
          "200": {
            "description": "成功返回渠道列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ChannelInfo" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["渠道管理"],
        "summary": "创建渠道",
        "description": "创建一个新的 AI 渠道。创建成功后返回完整的渠道信息。\n\n### 请求示例\n```json\n{\n  \"name\": \"gpt-4o\",\n  \"api_standard\": \"openai\",\n  \"price_input\": \"0.01\",\n  \"price_cached\": \"0.005\",\n  \"price_output\": \"0.03\",\n  \"rate_input\": \"1.0\",\n  \"rate_cached\": \"1.0\",\n  \"rate_output\": \"1.0\",\n  \"upstream_url\": \"https://api.openai.com/v1\",\n  \"api_key\": \"sk-xxx\"\n}\n```",
        "operationId": "createChannel",
        "security": [{ "ApiTokenAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "upstream_url"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "渠道名称（也作为 AI 调用时的模型名）",
                    "example": "deepseek-chat",
                    "minLength": 1
                  },
                  "api_standard": {
                    "type": "string",
                    "description": "API 协议标准",
                    "default": "openai",
                    "example": "openai"
                  },
                  "price_input": {
                    "type": "string",
                    "description": "输入单价（元/百万Token）",
                    "default": "0",
                    "example": "10"
                  },
                  "price_cached": {
                    "type": "string",
                    "description": "缓存命中单价（元/百万Token）",
                    "default": "0",
                    "example": "5"
                  },
                  "price_output": {
                    "type": "string",
                    "description": "输出单价（元/百万Token）",
                    "default": "0",
                    "example": "30"
                  },
                  "rate_input": {
                    "type": "string",
                    "description": "输入 Token 汇率（将上游 Token 数转换为标准 Token 数）",
                    "default": "1.0",
                    "example": "1.0"
                  },
                  "rate_cached": {
                    "type": "string",
                    "description": "缓存 Token 汇率",
                    "default": "1.0",
                    "example": "1.0"
                  },
                  "rate_output": {
                    "type": "string",
                    "description": "输出 Token 汇率",
                    "default": "1.0",
                    "example": "1.0"
                  },
                  "upstream_url": {
                    "type": "string",
                    "description": "上游 API 地址",
                    "example": "https://api.deepseek.com/v1",
                    "minLength": 1
                  },
                  "api_key": {
                    "type": "string",
                    "description": "上游 API 密钥",
                    "default": "",
                    "example": "sk-xxx"
                  },
                  "channel_type": {
                    "type": "string",
                    "description": "渠道厂商类型（deepseek/volcengine/qwen/baidu/tencent/openai/anthropic/google/custom）",
                    "default": "custom",
                    "example": "deepseek"
                  },
                  "price_reference_url": {
                    "type": "string",
                    "description": "计费价格参考链接",
                    "default": "",
                    "example": "https://api-docs.deepseek.com/zh-cn/quick_start/pricing"
                  },
                  "model": {
                    "type": "string",
                    "description": "上游实际调用的模型名（为空则使用渠道名）",
                    "default": "",
                    "example": "deepseek-chat"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功，返回新创建的渠道信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/ChannelInfo" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/channels/{id}": {
      "put": {
        "tags": ["渠道管理"],
        "summary": "更新渠道",
        "description": "更新指定渠道的配置信息。只需传入需要更新的字段，未传入的字段保持不变。\n\n### 请求示例\n```json\n{\n  \"price_input\": \"0.02\",\n  \"price_output\": \"0.04\"\n}\n```",
        "operationId": "updateChannel",
        "security": [{ "ApiTokenAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "渠道 ID",
            "schema": { "type": "integer", "format": "int64", "example": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "需要更新的字段，所有字段均为可选",
                "properties": {
                  "api_standard": {
                    "type": "string",
                    "description": "API 协议标准",
                    "example": "openai"
                  },
                  "price_input": {
                    "type": "string",
                    "description": "输入单价（元/百万Token）",
                    "example": "10"
                  },
                  "price_cached": {
                    "type": "string",
                    "description": "缓存命中单价（元/百万Token）",
                    "example": "5"
                  },
                  "price_output": {
                    "type": "string",
                    "description": "输出单价（元/百万Token）",
                    "example": "30"
                  },
                  "rate_input": {
                    "type": "string",
                    "description": "输入 Token 汇率",
                    "example": "1.5"
                  },
                  "rate_cached": {
                    "type": "string",
                    "description": "缓存 Token 汇率",
                    "example": "1.0"
                  },
                  "rate_output": {
                    "type": "string",
                    "description": "输出 Token 汇率",
                    "example": "1.5"
                  },
                  "upstream_url": {
                    "type": "string",
                    "description": "上游 API 地址",
                    "example": "https://api.deepseek.com/v1"
                  },
                  "api_key": {
                    "type": "string",
                    "description": "上游 API 密钥",
                    "example": "sk-xxx"
                  },
                  "channel_type": {
                    "type": "string",
                    "description": "渠道厂商类型（deepseek/volcengine/qwen/baidu/tencent/openai/anthropic/google/custom）",
                    "example": "deepseek"
                  },
                  "price_reference_url": {
                    "type": "string",
                    "description": "计费价格参考链接",
                    "example": "https://api-docs.deepseek.com/zh-cn/quick_start/pricing"
                  },
                  "model": {
                    "type": "string",
                    "description": "上游实际调用的模型名（为空则使用渠道名）",
                    "example": "deepseek-chat"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功，返回更新后的渠道信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/ChannelInfo" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["渠道管理"],
        "summary": "删除渠道",
        "description": "删除指定的渠道。删除后该渠道将无法用于 AI 调用。",
        "operationId": "deleteChannel",
        "security": [{ "ApiTokenAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "要删除的渠道 ID",
            "schema": { "type": "integer", "format": "int64", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deleted": { "type": "boolean", "example": true },
                        "id": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "AI 代理鉴权。使用系统生成的 Token（UUID 格式）。\n\n请求头格式：\n```\nAuthorization: Bearer {token}\n```\n\nToken 通过管理 API（POST /api/token/generate）生成。"
      },
      "ApiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "UUID",
        "description": "管理 API 鉴权。使用配置的 api_token（UUID 格式）。\n\n请求头格式：\n```\nAuthorization: Bearer {api_token}\n```\n\napi_token 在 config.toml 的 [admin] 节中配置。\n\n此 Token 可用于所有 /api/* 和 /admin/*/api 接口。"
      }
    },
    "schemas": {
      "ChannelInfo": {
        "type": "object",
        "description": "渠道信息",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "渠道 ID", "example": 1 },
          "name": { "type": "string", "description": "渠道名称（用于匹配 AI 请求中的 model 参数）", "example": "gpt-4o" },
          "api_standard": { "type": "string", "description": "API 协议标准", "example": "openai" },
          "price_input": { "type": "string", "description": "输入单价（元/百万Token）", "example": "10" },
          "price_cached": { "type": "string", "description": "缓存命中单价（元/百万Token）", "example": "5" },
          "price_output": { "type": "string", "description": "输出单价（元/百万Token）", "example": "30" },
          "rate_input": { "type": "string", "description": "输入 Token 汇率", "example": "1.0" },
          "rate_cached": { "type": "string", "description": "缓存 Token 汇率", "example": "1.0" },
          "rate_output": { "type": "string", "description": "输出 Token 汇率", "example": "1.0" },
          "upstream_url": { "type": "string", "description": "上游 API 地址", "example": "https://api.deepseek.com/v1" },
          "api_key": { "type": "string", "description": "上游 API 密钥", "example": "sk-xxx" },
          "channel_type": { "type": "string", "description": "渠道厂商类型（deepseek/volcengine/qwen/baidu/tencent/openai/anthropic/google/custom）", "example": "deepseek" },
          "price_reference_url": { "type": "string", "description": "计费价格参考链接", "example": "https://api-docs.deepseek.com/zh-cn/quick_start/pricing" },
          "model": { "type": "string", "description": "上游实际调用的模型名（为空则使用渠道名）", "example": "deepseek-chat" },
          "created_at": { "type": "string", "format": "date-time", "description": "创建时间", "example": "2025-01-01T00:00:00" },
          "updated_at": { "type": "string", "format": "date-time", "description": "更新时间", "example": "2025-01-01T00:00:00" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "错误响应",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "data": { "type": "null" },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "integer", "example": 401 },
              "message": { "type": "string", "example": "鉴权失败" }
            }
          },
          "message": { "type": "string", "example": "鉴权失败" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "鉴权失败 — Token 无效或未提供",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}